Refactor time functions test out of giant ZetaSqlDialectSpecTest
