+++ sdks/java/extensions/sql/zetasql/src/test/java/org/apache/beam/sdk/extensions/sql/zetasql/ZetaSqlDialectSpecTest.java
@@ -17,12 +17,6 @@
  */
 package org.apache.beam.sdk.extensions.sql.zetasql;
 
-import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseDate;
-import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseDateToValue;
-import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseTime;
-import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseTimeToValue;
-import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseTimestampWithTZToValue;
-import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseTimestampWithTimeZone;
 import static org.apache.beam.sdk.extensions.sql.zetasql.DateTimeUtils.parseTimestampWithUTCTimeZone;
 import static org.apache.beam.sdk.schemas.Schema.FieldType.DATETIME;
 import static org.junit.Assert.assertTrue;
@@ -70,17 +64,16 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** ZetaSQLDialectSpecTest. */
+/** ZetaSqlDialectSpecTest. */
 @RunWith(JUnit4.class)
-public class ZetaSQLDialectSpecTest extends ZetaSQLTestBase {
+public class ZetaSqlDialectSpecTest extends ZetaSqlTestBase {
 
   @Rule public transient TestPipeline pipeline = TestPipeline.create();
   @Rule public ExpectedException thrown = ExpectedException.none();
 
   @Before
   public void setUp() {
-    initializeBeamTableProvider();
-    initializeCalciteEnvironment();
+    initialize();
   }
 
   @Test
@@ -2181,692 +2174,6 @@ public class ZetaSQLDialectSpecTest extends ZetaSQLTestBase {
     pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
   }
 
-  /////////////////////////////////////////////////////////////////////////////
-  // DATE type tests
-  /////////////////////////////////////////////////////////////////////////////
-
-  @Test
-  public void testDateLiteral() {
-    String sql = "SELECT DATE '2020-3-30'";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_date", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2020, 3, 30))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateColumn() {
-    String sql = "SELECT FORMAT_DATE('%b-%d-%Y', date_field) FROM table_with_date";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addStringField("f_date_str").build())
-                .addValues("Dec-25-2008")
-                .build(),
-            Row.withSchema(Schema.builder().addStringField("f_date_str").build())
-                .addValues("Apr-07-2020")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testGroupByDate() {
-    String sql = "SELECT date_field, COUNT(*) FROM table_with_date GROUP BY date_field";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema =
-        Schema.builder()
-            .addLogicalTypeField("date_field", SqlTypes.DATE)
-            .addInt64Field("count")
-            .build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema).addValues(LocalDate.of(2008, 12, 25), 1L).build(),
-            Row.withSchema(schema).addValues(LocalDate.of(2020, 4, 7), 1L).build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testAggregateOnDate() {
-    String sql = "SELECT MAX(date_field) FROM table_with_date GROUP BY str_field";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder().addLogicalTypeField("date_field", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2020, 4, 7))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  // TODO[BEAM-9166]: Add a test for CURRENT_DATE function ("SELECT CURRENT_DATE()")
-
-  @Test
-  public void testExtractDate() {
-    String sql =
-        "WITH Dates AS (\n"
-            + "  SELECT DATE '2015-12-31' AS date UNION ALL\n"
-            + "  SELECT DATE '2016-01-01'\n"
-            + ")\n"
-            + "SELECT\n"
-            + "  EXTRACT(ISOYEAR FROM date) AS isoyear,\n"
-            + "  EXTRACT(YEAR FROM date) AS year,\n"
-            + "  EXTRACT(ISOWEEK FROM date) AS isoweek,\n"
-            // TODO[BEAM-9178]: Add tests for DATE_TRUNC and EXTRACT with "week with weekday" date
-            //  parts once they are supported
-            // + "  EXTRACT(WEEK FROM date) AS week,\n"
-            + "  EXTRACT(MONTH FROM date) AS month\n"
-            + "FROM Dates\n";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema =
-        Schema.builder()
-            .addField("isoyear", FieldType.INT64)
-            .addField("year", FieldType.INT64)
-            .addField("isoweek", FieldType.INT64)
-            // .addField("week", FieldType.INT64)
-            .addField("month", FieldType.INT64)
-            .build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema).addValues(2015L, 2015L, 53L /* , 52L */, 12L).build(),
-            Row.withSchema(schema).addValues(2015L, 2016L, 53L /* , 0L */, 1L).build());
-
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateFromYearMonthDay() {
-    String sql = "SELECT DATE(2008, 12, 25)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_date", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2008, 12, 25))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateFromTimestamp() {
-    String sql = "SELECT DATE(TIMESTAMP '2016-12-25 05:30:00+07', 'America/Los_Angeles')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_date", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2016, 12, 24))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateAdd() {
-    String sql =
-        "SELECT "
-            + "DATE_ADD(DATE '2008-12-25', INTERVAL 5 DAY), "
-            + "DATE_ADD(DATE '2008-12-25', INTERVAL 1 MONTH), "
-            + "DATE_ADD(DATE '2008-12-25', INTERVAL 1 YEAR), ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addLogicalTypeField("f_date1", SqlTypes.DATE)
-                        .addLogicalTypeField("f_date2", SqlTypes.DATE)
-                        .addLogicalTypeField("f_date3", SqlTypes.DATE)
-                        .build())
-                .addValues(
-                    LocalDate.of(2008, 12, 30),
-                    LocalDate.of(2009, 1, 25),
-                    LocalDate.of(2009, 12, 25))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateSub() {
-    String sql =
-        "SELECT "
-            + "DATE_SUB(DATE '2008-12-25', INTERVAL 5 DAY), "
-            + "DATE_SUB(DATE '2008-12-25', INTERVAL 1 MONTH), "
-            + "DATE_SUB(DATE '2008-12-25', INTERVAL 1 YEAR), ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addLogicalTypeField("f_date1", SqlTypes.DATE)
-                        .addLogicalTypeField("f_date2", SqlTypes.DATE)
-                        .addLogicalTypeField("f_date3", SqlTypes.DATE)
-                        .build())
-                .addValues(
-                    LocalDate.of(2008, 12, 20),
-                    LocalDate.of(2008, 11, 25),
-                    LocalDate.of(2007, 12, 25))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateDiff() {
-    String sql = "SELECT DATE_DIFF(DATE '2010-07-07', DATE '2008-12-25', DAY)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_date_diff").build())
-                .addValues(559L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateDiffNegativeResult() {
-    String sql = "SELECT DATE_DIFF(DATE '2017-12-17', DATE '2017-12-18', ISOWEEK)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_date_diff").build())
-                .addValues(-1L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateTrunc() {
-    String sql = "SELECT DATE_TRUNC(DATE '2015-06-15', ISOYEAR)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder().addLogicalTypeField("f_date_trunc", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2014, 12, 29))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testFormatDate() {
-    String sql = "SELECT FORMAT_DATE('%b-%d-%Y', DATE '2008-12-25')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addStringField("f_date_str").build())
-                .addValues("Dec-25-2008")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testParseDate() {
-    String sql = "SELECT PARSE_DATE('%m %d %y', '10 14 18')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_date", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2018, 10, 14))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateToUnixInt64() {
-    String sql = "SELECT UNIX_DATE(DATE '2008-12-25')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_unix_date").build())
-                .addValues(14238L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testDateFromUnixInt64() {
-    String sql = "SELECT DATE_FROM_UNIX_DATE(14238)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_date", SqlTypes.DATE).build())
-                .addValues(LocalDate.of(2008, 12, 25))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  /////////////////////////////////////////////////////////////////////////////
-  // TIME type tests
-  /////////////////////////////////////////////////////////////////////////////
-
-  @Test
-  public void testTimeLiteral() {
-    String sql = "SELECT TIME '15:30:00', TIME '15:30:00.135246' ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addLogicalTypeField("f_time1", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time2", SqlTypes.TIME)
-                        .build())
-                .addValues(LocalTime.of(15, 30, 0))
-                .addValues(LocalTime.of(15, 30, 0, 135246000))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeColumn() {
-    String sql = "SELECT FORMAT_TIME('%T', time_field) FROM table_with_time";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addStringField("f_time_str").build())
-                .addValues("15:30:00")
-                .build(),
-            Row.withSchema(Schema.builder().addStringField("f_time_str").build())
-                .addValues("23:35:59")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testGroupByTime() {
-    String sql = "SELECT time_field, COUNT(*) FROM table_with_time GROUP BY time_field";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema =
-        Schema.builder()
-            .addLogicalTypeField("time_field", SqlTypes.TIME)
-            .addInt64Field("count")
-            .build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema).addValues(LocalTime.of(15, 30, 0), 1L).build(),
-            Row.withSchema(schema).addValues(LocalTime.of(23, 35, 59), 1L).build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testAggregateOnTime() {
-    String sql = "SELECT MAX(time_field) FROM table_with_time GROUP BY str_field";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder().addLogicalTypeField("time_field", SqlTypes.TIME).build())
-                .addValues(LocalTime.of(23, 35, 59))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  // TODO[BEAM-9166]: Add a test for CURRENT_TIME function ("SELECT CURRENT_TIME()")
-
-  @Test
-  public void testExtractTime() {
-    String sql =
-        "SELECT "
-            + "EXTRACT(HOUR FROM TIME '15:30:35.123456') as hour, "
-            + "EXTRACT(MINUTE FROM TIME '15:30:35.123456') as minute, "
-            + "EXTRACT(SECOND FROM TIME '15:30:35.123456') as second, "
-            + "EXTRACT(MILLISECOND FROM TIME '15:30:35.123456') as millisecond, "
-            + "EXTRACT(MICROSECOND FROM TIME '15:30:35.123456') as microsecond ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema =
-        Schema.builder()
-            .addField("hour", FieldType.INT64)
-            .addField("minute", FieldType.INT64)
-            .addField("second", FieldType.INT64)
-            .addField("millisecond", FieldType.INT64)
-            .addField("microsecond", FieldType.INT64)
-            .build();
-    PAssert.that(stream)
-        .containsInAnyOrder(Row.withSchema(schema).addValues(15L, 30L, 35L, 123L, 123456L).build());
-
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeFromHourMinuteSecond() {
-    String sql = "SELECT TIME(15, 30, 0)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_time", SqlTypes.TIME).build())
-                .addValues(LocalTime.of(15, 30, 0))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeFromTimestamp() {
-    String sql = "SELECT TIME(TIMESTAMP '2008-12-25 15:30:00+08', 'America/Los_Angeles')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_time", SqlTypes.TIME).build())
-                .addValues(LocalTime.of(23, 30, 0))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeAdd() {
-    String sql =
-        "SELECT "
-            + "TIME_ADD(TIME '15:30:00', INTERVAL 10 MICROSECOND), "
-            + "TIME_ADD(TIME '15:30:00', INTERVAL 10 MILLISECOND), "
-            + "TIME_ADD(TIME '15:30:00', INTERVAL 10 SECOND), "
-            + "TIME_ADD(TIME '15:30:00', INTERVAL 10 MINUTE), "
-            + "TIME_ADD(TIME '15:30:00', INTERVAL 10 HOUR) ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addLogicalTypeField("f_time1", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time2", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time3", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time4", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time5", SqlTypes.TIME)
-                        .build())
-                .addValues(
-                    LocalTime.of(15, 30, 0, 10000),
-                    LocalTime.of(15, 30, 0, 10000000),
-                    LocalTime.of(15, 30, 10, 0),
-                    LocalTime.of(15, 40, 0, 0),
-                    LocalTime.of(1, 30, 0, 0))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeSub() {
-    String sql =
-        "SELECT "
-            + "TIME_SUB(TIME '15:30:00', INTERVAL 10 MICROSECOND), "
-            + "TIME_SUB(TIME '15:30:00', INTERVAL 10 MILLISECOND), "
-            + "TIME_SUB(TIME '15:30:00', INTERVAL 10 SECOND), "
-            + "TIME_SUB(TIME '15:30:00', INTERVAL 10 MINUTE), "
-            + "TIME_SUB(TIME '15:30:00', INTERVAL 10 HOUR) ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addLogicalTypeField("f_time1", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time2", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time3", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time4", SqlTypes.TIME)
-                        .addLogicalTypeField("f_time5", SqlTypes.TIME)
-                        .build())
-                .addValues(
-                    LocalTime.of(15, 29, 59, 999990000),
-                    LocalTime.of(15, 29, 59, 990000000),
-                    LocalTime.of(15, 29, 50, 0),
-                    LocalTime.of(15, 20, 0, 0),
-                    LocalTime.of(5, 30, 0, 0))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeDiff() {
-    String sql = "SELECT TIME_DIFF(TIME '15:30:00', TIME '14:35:00', MINUTE)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_time_diff").build())
-                .addValues(55L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeDiffNegativeResult() {
-    String sql = "SELECT TIME_DIFF(TIME '14:35:00', TIME '15:30:00', MINUTE)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_time_diff").build())
-                .addValues(-55L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimeTrunc() {
-    String sql = "SELECT TIME_TRUNC(TIME '15:30:35', HOUR)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder().addLogicalTypeField("f_time_trunc", SqlTypes.TIME).build())
-                .addValues(LocalTime.of(15, 0, 0))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testFormatTime() {
-    String sql = "SELECT FORMAT_TIME('%R', TIME '15:30:00')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addStringField("f_time_str").build())
-                .addValues("15:30")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testParseTime() {
-    String sql = "SELECT PARSE_TIME('%H', '15')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addLogicalTypeField("f_time", SqlTypes.TIME).build())
-                .addValues(LocalTime.of(15, 0, 0))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  @Ignore("https://jira.apache.org/jira/browse/BEAM-10340")
-  public void testCastBetweenTimeAndString() {
-    String sql =
-        "SELECT CAST(s1 as TIME) as t2, CAST(t1 as STRING) as s2 FROM "
-            + "(SELECT '12:34:56.123456' as s1, TIME '12:34:56.123456' as t1)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addLogicalTypeField("t2", SqlTypes.TIME)
-                        .addStringField("s2")
-                        .build())
-                .addValues(LocalTime.of(12, 34, 56, 123456000), "12:34:56.123456")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  /////////////////////////////////////////////////////////////////////////////
-  // TIMESTAMP type tests
-  /////////////////////////////////////////////////////////////////////////////
-
-  @Test
-  public void testTimestampMicrosecondUnsupported() {
-    String sql =
-        "WITH Timestamps AS (\n"
-            + "  SELECT TIMESTAMP '2000-01-01 00:11:22.345678+00' as timestamp\n"
-            + ")\n"
-            + "SELECT\n"
-            + "  timestamp,\n"
-            + "  EXTRACT(ISOYEAR FROM timestamp) AS isoyear,\n"
-            + "  EXTRACT(YEAR FROM timestamp) AS year,\n"
-            + "  EXTRACT(ISOWEEK FROM timestamp) AS week,\n"
-            + "  EXTRACT(MINUTE FROM timestamp) AS minute\n"
-            + "FROM Timestamps\n";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    thrown.expect(UnsupportedOperationException.class);
-    zetaSQLQueryPlanner.convertToBeamRel(sql);
-  }
-
-  @Test
-  public void testTimestampLiteralWithoutTimeZone() {
-    String sql = "SELECT TIMESTAMP '2016-12-25 05:30:00'";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("field1").build())
-                .addValues(parseTimestampWithUTCTimeZone("2016-12-25 05:30:00"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampLiteralWithUTCTimeZone() {
-    String sql = "SELECT TIMESTAMP '2016-12-25 05:30:00+00'";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("field1").build())
-                .addValues(parseTimestampWithUTCTimeZone("2016-12-25 05:30:00"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
   @Test
   public void testMultipleSelectStatementsThrowsException() {
     String sql = "SELECT 1; SELECT 2;";
@@ -3040,369 +2347,6 @@ public class ZetaSQLDialectSpecTest extends ZetaSQLTestBase {
     thrown.expect(SqlException.class);
     thrown.expectMessage("Table-valued function not found: FunctionNotFound");
     zetaSQLQueryPlanner.convertToBeamRel(sql);
-  }
-
-  @Test
-  public void testTimestampLiteralWithNonUTCTimeZone() {
-    String sql = "SELECT TIMESTAMP '2018-12-10 10:38:59-10:00'";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("f_timestamp_with_time_zone").build())
-                .addValues(parseTimestampWithTimeZone("2018-12-10 10:38:59-1000"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  // TODO[BEAM-9166]: Add a test for CURRENT_TIMESTAMP function ("SELECT CURRENT_TIMESTAMP()")
-
-  @Test
-  public void testExtractTimestamp() {
-    String sql =
-        "WITH Timestamps AS (\n"
-            + "  SELECT TIMESTAMP '2007-12-31 12:34:56' AS timestamp UNION ALL\n"
-            + "  SELECT TIMESTAMP '2009-12-31'\n"
-            + ")\n"
-            + "SELECT\n"
-            + "  EXTRACT(ISOYEAR FROM timestamp) AS isoyear,\n"
-            + "  EXTRACT(YEAR FROM timestamp) AS year,\n"
-            + "  EXTRACT(ISOWEEK FROM timestamp) AS isoweek,\n"
-            // TODO[BEAM-9178]: Add tests for TIMESTAMP_TRUNC and EXTRACT with "week with weekday"
-            //  date parts once they are supported
-            // + "  EXTRACT(WEEK FROM timestamp) AS week,\n"
-            + "  EXTRACT(MINUTE FROM timestamp) AS minute\n"
-            + "FROM Timestamps\n";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema =
-        Schema.builder()
-            .addField("isoyear", FieldType.INT64)
-            .addField("year", FieldType.INT64)
-            .addField("isoweek", FieldType.INT64)
-            // .addField("week", FieldType.INT64)
-            .addField("minute", FieldType.INT64)
-            .build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema).addValues(2008L, 2007L, 1L /* , 53L */, 34L).build(),
-            Row.withSchema(schema).addValues(2009L, 2009L, 53L /* , 52L */, 0L).build());
-
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testExtractTimestampAtTimeZoneUnsupported() {
-    String sql =
-        "WITH Timestamps AS (\n"
-            + "  SELECT TIMESTAMP '2017-05-26' AS timestamp\n"
-            + ")\n"
-            + "SELECT\n"
-            + "  timestamp,\n"
-            + "  EXTRACT(HOUR FROM timestamp AT TIME ZONE 'America/Vancouver') AS hour,\n"
-            + "  EXTRACT(DAY FROM timestamp AT TIME ZONE 'America/Vancouver') AS day\n"
-            + "FROM Timestamps\n";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    thrown.expect(UnsupportedOperationException.class);
-    zetaSQLQueryPlanner.convertToBeamRel(sql);
-  }
-
-  @Test
-  public void testExtractDateFromTimestampUnsupported() {
-    String sql =
-        "WITH Timestamps AS (\n"
-            + "  SELECT TIMESTAMP '2017-05-26' AS ts\n"
-            + ")\n"
-            + "SELECT\n"
-            + "  ts,\n"
-            + "  EXTRACT(DATE FROM ts) AS dt\n"
-            + "FROM Timestamps\n";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    thrown.expect(SqlException.class);
-    zetaSQLQueryPlanner.convertToBeamRel(sql);
-  }
-
-  @Test
-  public void testStringFromTimestamp() {
-    String sql = "SELECT STRING(TIMESTAMP '2008-12-25 15:30:00', 'America/Los_Angeles')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addStringField("f_timestamp_string").build())
-                .addValues("2008-12-25 07:30:00-08")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampFromString() {
-    String sql = "SELECT TIMESTAMP('2008-12-25 15:30:00', 'America/Los_Angeles')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("f_timestamp").build())
-                .addValues(parseTimestampWithTimeZone("2008-12-25 15:30:00-08"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampFromDate() {
-    String sql = "SELECT TIMESTAMP(DATE '2014-01-31')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("f_timestamp").build())
-                .addValues(parseTimestampWithTimeZone("2014-01-31 00:00:00+00"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  // test default timezone works properly in query execution stage
-  public void testTimestampFromDateWithDefaultTimezoneSet() {
-    String sql = "SELECT TIMESTAMP(DATE '2014-01-31')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    zetaSQLQueryPlanner.setDefaultTimezone("Asia/Shanghai");
-    pipeline
-        .getOptions()
-        .as(BeamSqlPipelineOptions.class)
-        .setZetaSqlDefaultTimezone("Asia/Shanghai");
-
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("f_timestamp").build())
-                .addValues(parseTimestampWithTimeZone("2014-01-31 00:00:00+08"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampAdd() {
-    String sql =
-        "SELECT "
-            + "TIMESTAMP_ADD(TIMESTAMP '2008-12-25 15:30:00 UTC', INTERVAL 5+5 MINUTE), "
-            + "TIMESTAMP_ADD(TIMESTAMP '2008-12-25 15:30:00+07:30', INTERVAL 10 MINUTE)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addDateTimeField("f_timestamp_add")
-                        .addDateTimeField("f_timestamp_with_time_zone_add")
-                        .build())
-                .addValues(
-                    DateTimeUtils.parseTimestampWithUTCTimeZone("2008-12-25 15:40:00"),
-                    parseTimestampWithTimeZone("2008-12-25 15:40:00+0730"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampSub() {
-    String sql =
-        "SELECT "
-            + "TIMESTAMP_SUB(TIMESTAMP '2008-12-25 15:30:00 UTC', INTERVAL 5+5 MINUTE), "
-            + "TIMESTAMP_SUB(TIMESTAMP '2008-12-25 15:30:00+07:30', INTERVAL 10 MINUTE)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addDateTimeField("f_timestamp_sub")
-                        .addDateTimeField("f_timestamp_with_time_zone_sub")
-                        .build())
-                .addValues(
-                    DateTimeUtils.parseTimestampWithUTCTimeZone("2008-12-25 15:20:00"),
-                    parseTimestampWithTimeZone("2008-12-25 15:20:00+0730"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampDiff() {
-    String sql =
-        "SELECT TIMESTAMP_DIFF("
-            + "TIMESTAMP '2018-10-14 15:30:00.000 UTC', "
-            + "TIMESTAMP '2018-08-14 15:05:00.001 UTC', "
-            + "MILLISECOND)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_timestamp_diff").build())
-                .addValues((61L * 24 * 60 + 25) * 60 * 1000 - 1)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampDiffNegativeResult() {
-    String sql = "SELECT TIMESTAMP_DIFF(TIMESTAMP '2018-08-14', TIMESTAMP '2018-10-14', DAY)";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addInt64Field("f_timestamp_diff").build())
-                .addValues(-61L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampTrunc() {
-    String sql = "SELECT TIMESTAMP_TRUNC(TIMESTAMP '2017-11-06 00:00:00+12', ISOWEEK, 'UTC')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("f_timestamp_trunc").build())
-                .addValues(DateTimeUtils.parseTimestampWithUTCTimeZone("2017-10-30 00:00:00"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testFormatTimestamp() {
-    String sql = "SELECT FORMAT_TIMESTAMP('%D %T', TIMESTAMP '2018-10-14 15:30:00.123+00', 'UTC')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addStringField("f_timestamp_str").build())
-                .addValues("10/14/18 15:30:00")
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testParseTimestamp() {
-    String sql = "SELECT PARSE_TIMESTAMP('%m-%d-%y %T', '10-14-18 15:30:00', 'UTC')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(Schema.builder().addDateTimeField("f_timestamp").build())
-                .addValues(DateTimeUtils.parseTimestampWithUTCTimeZone("2018-10-14 15:30:00"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampFromInt64() {
-    String sql = "SELECT TIMESTAMP_SECONDS(1230219000), TIMESTAMP_MILLIS(1230219000123) ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addDateTimeField("f_timestamp_seconds")
-                        .addDateTimeField("f_timestamp_millis")
-                        .build())
-                .addValues(
-                    DateTimeUtils.parseTimestampWithUTCTimeZone("2008-12-25 15:30:00"),
-                    DateTimeUtils.parseTimestampWithUTCTimeZone("2008-12-25 15:30:00.123"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampToUnixInt64() {
-    String sql =
-        "SELECT "
-            + "UNIX_SECONDS(TIMESTAMP '2008-12-25 15:30:00 UTC'), "
-            + "UNIX_MILLIS(TIMESTAMP '2008-12-25 15:30:00.123 UTC')";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addInt64Field("f_unix_seconds")
-                        .addInt64Field("f_unix_millis")
-                        .build())
-                .addValues(1230219000L, 1230219000123L)
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampFromUnixInt64() {
-    String sql =
-        "SELECT "
-            + "TIMESTAMP_FROM_UNIX_SECONDS(1230219000), "
-            + "TIMESTAMP_FROM_UNIX_MILLIS(1230219000123) ";
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(
-                    Schema.builder()
-                        .addDateTimeField("f_timestamp_seconds")
-                        .addDateTimeField("f_timestamp_millis")
-                        .build())
-                .addValues(
-                    DateTimeUtils.parseTimestampWithUTCTimeZone("2008-12-25 15:30:00"),
-                    DateTimeUtils.parseTimestampWithUTCTimeZone("2008-12-25 15:30:00.123"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
   }
 
   @Test
@@ -4129,16 +3073,6 @@ public class ZetaSQLDialectSpecTest extends ZetaSQLTestBase {
   }
 
   @Test
-  @Ignore("Does not support DateTime literal.")
-  public void testDateTimeLiteral() {
-    String sql = "SELECT DATETIME '2018-01-01 05:30:00.334'";
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    thrown.expect(RuntimeException.class);
-    thrown.expectMessage("Unsupported ResolvedLiteral type: DATETIME");
-    zetaSQLQueryPlanner.convertToBeamRel(sql);
-  }
-
-  @Test
   public void testConcatWithOneParameters() {
     String sql = "SELECT concat('abc')";
     ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
@@ -4693,6 +3627,29 @@ public class ZetaSQLDialectSpecTest extends ZetaSQLTestBase {
   }
 
   @Test
+  @Ignore("https://jira.apache.org/jira/browse/BEAM-10340")
+  public void testCastBetweenTimeAndString() {
+    String sql =
+        "SELECT CAST(s1 as TIME) as t2, CAST(t1 as STRING) as s2 FROM "
+            + "(SELECT '12:34:56.123456' as s1, TIME '12:34:56.123456' as t1)";
+
+    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
+    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql);
+    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
+
+    PAssert.that(stream)
+        .containsInAnyOrder(
+            Row.withSchema(
+                    Schema.builder()
+                        .addLogicalTypeField("t2", SqlTypes.TIME)
+                        .addStringField("s2")
+                        .build())
+                .addValues(LocalTime.of(12, 34, 56, 123456000), "12:34:56.123456")
+                .build());
+    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
+  }
+
+  @Test
   public void testCastStringToString() {
     String sql = "SELECT CAST(@p0 AS STRING)";
     ImmutableMap<String, Value> params = ImmutableMap.of("p0", Value.createStringValue(""));
@@ -4730,117 +3687,6 @@ public class ZetaSQLDialectSpecTest extends ZetaSQLTestBase {
     final Schema schema = Schema.builder().addStringField("field1").build();
 
     PAssert.that(stream).containsInAnyOrder(Row.withSchema(schema).addValues("hi").build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  @Ignore("Does not support DATE_ADD.")
-  public void testDateAddWithParameter() {
-    String sql =
-        "SELECT "
-            + "DATE_ADD(@p0, INTERVAL @p1 DAY), "
-            + "DATE_ADD(@p2, INTERVAL @p3 DAY), "
-            + "DATE_ADD(@p4, INTERVAL @p5 YEAR), "
-            + "DATE_ADD(@p6, INTERVAL @p7 DAY), "
-            + "DATE_ADD(@p8, INTERVAL @p9 MONTH)";
-    // Value
-    ImmutableMap<String, Value> params =
-        ImmutableMap.<String, Value>builder()
-            .put("p0", Value.createDateValue(0)) // 1970-01-01
-            .put("p1", Value.createInt64Value(2L))
-            .put("p2", parseDateToValue("2019-01-01"))
-            .put("p3", Value.createInt64Value(2L))
-            .put("p4", Value.createSimpleNullValue(TypeKind.TYPE_DATE))
-            .put("p5", Value.createInt64Value(1L))
-            .put("p6", parseDateToValue("2000-02-29"))
-            .put("p7", Value.createInt64Value(-365L))
-            .put("p8", parseDateToValue("1999-03-31"))
-            .put("p9", Value.createInt64Value(-1L))
-            .build();
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql, params);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema =
-        Schema.builder()
-            .addDateTimeField("field1")
-            .addDateTimeField("field2")
-            .addNullableField("field3", DATETIME)
-            .addDateTimeField("field4")
-            .addDateTimeField("field5")
-            .build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema)
-                .addValues(
-                    parseDate("1970-01-03"),
-                    parseDate("2019-01-03"),
-                    null,
-                    parseDate("1999-03-01"),
-                    parseDate("1999-02-28"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  @Ignore("Does not support TIME_ADD.")
-  public void testTimeAddWithParameter() {
-    String sql = "SELECT TIME_ADD(@p0, INTERVAL @p1 SECOND)";
-    ImmutableMap<String, Value> params =
-        ImmutableMap.of(
-            "p0", parseTimeToValue("12:13:14.123"),
-            "p1", Value.createInt64Value(1L));
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql, params);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema = Schema.builder().addDateTimeField("field1").build();
-    PAssert.that(stream)
-        .containsInAnyOrder(Row.withSchema(schema).addValues(parseTime("12:13:15.123")).build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampAddWithParameter1() {
-    String sql = "SELECT TIMESTAMP_ADD(@p0, INTERVAL @p1 MILLISECOND)";
-    ImmutableMap<String, Value> params =
-        ImmutableMap.of(
-            "p0", parseTimestampWithTZToValue("2001-01-01 00:00:00+00"),
-            "p1", Value.createInt64Value(1L));
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql, params);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema = Schema.builder().addDateTimeField("field1").build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema)
-                .addValues(parseTimestampWithTimeZone("2001-01-01 00:00:00.001+00"))
-                .build());
-    pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
-  }
-
-  @Test
-  public void testTimestampAddWithParameter2() {
-    String sql = "SELECT TIMESTAMP_ADD(@p0, INTERVAL @p1 MINUTE)";
-    ImmutableMap<String, Value> params =
-        ImmutableMap.of(
-            "p0", parseTimestampWithTZToValue("2008-12-25 15:30:00+07:30"),
-            "p1", Value.createInt64Value(10L));
-
-    ZetaSQLQueryPlanner zetaSQLQueryPlanner = new ZetaSQLQueryPlanner(config);
-    BeamRelNode beamRelNode = zetaSQLQueryPlanner.convertToBeamRel(sql, params);
-    PCollection<Row> stream = BeamSqlRelUtils.toPCollection(pipeline, beamRelNode);
-
-    final Schema schema = Schema.builder().addDateTimeField("field1").build();
-    PAssert.that(stream)
-        .containsInAnyOrder(
-            Row.withSchema(schema)
-                .addValues(parseTimestampWithTimeZone("2008-12-25 15:40:00+07:30"))
-                .build());
     pipeline.run().waitUntilFinish(Duration.standardMinutes(PIPELINE_EXECUTION_WAITTIME_MINUTES));
   }
